re factor code in smaller functions
