+++ README.md
@@ -22,7 +22,7 @@ This Hardhat plugin provides two tasks and two functions to help you manage and 
 
 To install the Hardhat Transaction-Retry-Tool Plugin, run the following command in your project directory:
 
-```shell
+```bash
 npm i transaction-retry-tool
 ```
 
@@ -36,7 +36,7 @@ And that's it! You can now use the two tasks and functions provided by this plug
 
 
 ## Directories
- - [src/](./src/)
+ - [.vscode/](./.vscode/) - [src/](./src/)
 
  - [.eslintrc.js](./.eslintrc.js)
  - [.npmignore](./.npmignore)
@@ -57,7 +57,7 @@ And that's it! You can now use the two tasks and functions provided by this plug
 
 The plugin adds two tasks to the Hardhat CLI:
 
-```shell
+```bash
 npx hardhat retry
 ```
 
@@ -112,7 +112,10 @@ transaction-retry-tool/
 │   tsconfig.json
 │   tsconfig.prod.json
 │   tslint.json
+└─── .vscode/
 └─── src/
+   │   README.md
+   │   settings.json
    │   README.md
    │   TransactionRetry.ts
    │   getGasCost.ts
@@ -120,5 +123,6 @@ transaction-retry-tool/
    │   retryTransaction.ts
    │   serveTasks.ts
    │   type-extensions.ts
+   │   utils.ts
 ```
 ## Don't hesitate to contribute to this project.
